requirements.txt
torch
torchvision
wandb
hydra-core
pyyaml
omegaconf
colorama
tqdm
opencv-python
matplotlib
pytorch_lightning
click
moviepy
imageio
einops
